requirements.txt
aniso8601==8.1.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
djangorestframework==3.12.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
pytz==2020.5
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
